server/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
# For local dev; for Postgres in cloud, use psycopg2-binary
psycopg2-binary==2.9.9
